desktop: redesign transaction details around a hero amount
Rework TxDetailsWindow as a sent or received hero, extracting the shared
hero, sheet and tx id builders from the withdraw popups into TxHeroWindow.
