+++ core/src/main/resources/i18n/displayStrings.properties
@@ -1111,7 +1111,6 @@ funds.withdrawal.senderAmount=Sender's amount
 funds.withdrawal.feeExcluded=Amount excludes mining fee
 funds.withdrawal.feeIncluded=Amount includes mining fee
 funds.withdrawal.fromLabel=Withdraw from address
-funds.withdrawal.memoLabel=Withdrawal memo
 funds.withdrawal.noFundsAvailable=No funds are available for withdrawal
 funds.withdrawal.confirm.headline=Confirm send
 funds.withdrawal.confirm.to=To
@@ -2102,11 +2101,9 @@ tradeDetailsWindow.agentAddresses=Arbitrator/Mediator
 tradeDetailsWindow.detailData=Detail data
 
 txDetailsWindow.headline=Transaction Details
-txDetailsWindow.xmr.noteSent=You have sent XMR.
-txDetailsWindow.xmr.noteReceived=You have received XMR.
+txDetailsWindow.received=Received
 txDetailsWindow.sentTo=Sent to
 txDetailsWindow.receivedWith=Received with
-txDetailsWindow.txId=TxId
 txDetailsWindow.txKey=Transaction Key
 
 closedTradesSummaryWindow.headline=Trade history summary
