+++ desktop/src/main/java/haveno/desktop/main/overlays/windows/WithdrawConfirmationWindow.java
@@ -18,32 +18,22 @@
 package haveno.desktop.main.overlays.windows;
 
 import de.jensd.fx.glyphs.materialdesignicons.MaterialDesignIcon;
-import haveno.common.UserThread;
-import haveno.common.util.Utilities;
 import haveno.core.locale.Res;
 import haveno.core.trade.HavenoUtils;
 import haveno.desktop.components.AutoTooltipButton;
 import haveno.desktop.components.AutoTooltipLabel;
-import haveno.desktop.main.overlays.Overlay;
-import haveno.desktop.util.GlyphsDude;
-import haveno.desktop.util.Layout;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
-import javafx.scene.Cursor;
-import javafx.scene.Node;
 import javafx.scene.control.Label;
-import javafx.scene.control.Tooltip;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.HBox;
-import javafx.scene.layout.Priority;
-import javafx.scene.layout.Region;
 import javafx.scene.layout.VBox;
 
 import java.math.BigInteger;
 import java.util.function.Function;
 
 /** Confirmation showing the amount, destination and fee of a pending withdrawal. */
-public class WithdrawConfirmationWindow extends Overlay<WithdrawConfirmationWindow> {
+public class WithdrawConfirmationWindow extends TxHeroWindow<WithdrawConfirmationWindow> {
 
     private final BigInteger amount;
     private final BigInteger fee;
@@ -51,7 +41,6 @@ public class WithdrawConfirmationWindow extends Overlay<WithdrawConfirmationWind
     private final Function<BigInteger, String> fiatText; // approximate fiat for an amount, or null while no price
 
     public WithdrawConfirmationWindow(BigInteger amount, BigInteger fee, String address, Function<BigInteger, String> fiatText) {
-        type = Type.Confirmation;
         this.amount = amount;
         this.fee = fee;
         this.address = address;
@@ -60,62 +49,12 @@ public class WithdrawConfirmationWindow extends Overlay<WithdrawConfirmationWind
 
     @Override
     public void show() {
-        width = 560;
-        createGridPane();
-        gridPane.setPadding(new Insets(76, 78, 74, 78)); // popup-bg reserves 44 per side for the dropshadow
-        addContent();
-        addButtons();
-        applyStyles();
-        display();
+        showHeroWindow();
     }
 
-    private void addContent() {
-        // hero: the amount being sent, front and center
-        Label icon = new Label();
-        icon.setGraphic(GlyphsDude.createIcon(MaterialDesignIcon.SEND, "1.6em"));
-        icon.getStyleClass().add("confirm-send-icon");
-        icon.setMinSize(50, 50);
-        icon.setMaxSize(50, 50);
-        icon.setAlignment(Pos.CENTER);
-
-        Label title = new AutoTooltipLabel(Res.get("funds.withdrawal.confirm.headline"));
-        title.getStyleClass().add("confirm-send-title");
-        Label amountLabel = new AutoTooltipLabel(HavenoUtils.formatXmr(amount, true));
-        amountLabel.getStyleClass().add("confirm-send-amount");
-
-        VBox hero = new VBox(icon, title, amountLabel);
-        hero.setAlignment(Pos.CENTER);
-        VBox.setMargin(title, new Insets(16, 0, 0, 0));
-        VBox.setMargin(amountLabel, new Insets(8, 0, 0, 0));
-        String amountFiat = fiatText.apply(amount);
-        if (amountFiat != null) {
-            Label fiatLabel = new AutoTooltipLabel(amountFiat);
-            fiatLabel.getStyleClass().add("confirm-send-fiat");
-            VBox.setMargin(fiatLabel, new Insets(3, 0, 0, 0));
-            hero.getChildren().add(fiatLabel);
-        }
-
-        // destination with a copy shortcut; the address wraps in full below
-        Label copyIcon = new Label();
-        copyIcon.setGraphic(GlyphsDude.createIcon(MaterialDesignIcon.CONTENT_COPY, "1.1em"));
-        copyIcon.getStyleClass().add("confirm-send-copy");
-        copyIcon.setCursor(Cursor.HAND);
-        copyIcon.setTooltip(new Tooltip(Res.get("shared.copyToClipboard")));
-        copyIcon.setOnMouseClicked(e -> {
-            Utilities.copyToClipboard(address);
-            Tooltip tp = new Tooltip(Res.get("shared.copiedToClipboard"));
-            tp.show(copyIcon, e.getScreenX() + Layout.PADDING, e.getScreenY() + Layout.PADDING);
-            UserThread.runAfter(tp::hide, 1);
-        });
-        Region toSpacer = new Region();
-        HBox.setHgrow(toSpacer, Priority.ALWAYS);
-        HBox toRow = new HBox(rowLabel(Res.get("funds.withdrawal.confirm.to")), toSpacer, copyIcon);
-        toRow.setAlignment(Pos.CENTER_LEFT);
-        Label addressLabel = new Label(address);
-        addressLabel.setWrapText(true);
-        addressLabel.getStyleClass().add("confirm-send-address");
-        VBox toGroup = new VBox(3, toRow, addressLabel);
-
+    @Override
+    protected void addContent() {
+        VBox toGroup = sheetGroup(Res.get("funds.withdrawal.confirm.to"), wrappedLabel(address, "confirm-send-address"), copyIcon(address));
         HBox feeRow = detailRow(Res.get("funds.withdrawal.confirm.networkFee"), valueLabel(HavenoUtils.formatXmr(fee, true)));
 
         // total debited from the wallet, with a fiat approximation
@@ -132,12 +71,9 @@ public class WithdrawConfirmationWindow extends Overlay<WithdrawConfirmationWind
         }
         HBox totalRow = detailRow(Res.get("funds.withdrawal.confirm.total"), totalBox);
 
-        VBox sheet = new VBox(14, toGroup, divider(), feeRow, divider(), totalRow);
-        sheet.getStyleClass().add("confirm-send-sheet");
-
-        VBox content = new VBox(24, hero, sheet);
-        content.setFillWidth(true);
-        gridPane.add(content, 0, ++rowIndex, 2, 1);
+        addHeroContent(
+                hero(MaterialDesignIcon.SEND, "confirm-send-icon", Res.get("funds.withdrawal.confirm.headline"), amount, fiatText.apply(amount)),
+                sheet(toGroup, feeRow, totalRow));
     }
 
     @Override
@@ -160,31 +96,5 @@ public class WithdrawConfirmationWindow extends Overlay<WithdrawConfirmationWind
         buttons.setFillWidth(true);
         gridPane.add(buttons, 0, ++rowIndex, 2, 1);
         GridPane.setMargin(buttons, new Insets(26, 0, 0, 0));
-    }
-
-    private static Label rowLabel(String text) {
-        Label label = new AutoTooltipLabel(text);
-        label.getStyleClass().add("confirm-send-row-label");
-        return label;
-    }
-
-    private static Label valueLabel(String text) {
-        Label label = new AutoTooltipLabel(text);
-        label.getStyleClass().add("confirm-send-row-value");
-        return label;
-    }
-
-    private static HBox detailRow(String labelText, Node value) {
-        Region spacer = new Region();
-        HBox.setHgrow(spacer, Priority.ALWAYS);
-        HBox row = new HBox(rowLabel(labelText), spacer, value);
-        row.setAlignment(Pos.CENTER_LEFT);
-        return row;
-    }
-
-    private static Region divider() {
-        Region line = new Region();
-        line.getStyleClass().add("confirm-send-divider");
-        return line;
     }
 }
