+++ desktop/src/main/java/haveno/desktop/util/GUIUtil.java
@@ -50,6 +50,8 @@ import haveno.core.locale.TradeCurrency;
 import haveno.core.payment.PaymentAccount;
 import haveno.core.payment.PaymentAccountList;
 import haveno.core.payment.payload.PaymentMethod;
+import haveno.core.provider.price.MarketPrice;
+import haveno.core.provider.price.PriceFeedService;
 import haveno.core.trade.HavenoUtils;
 import haveno.core.trade.Trade;
 import haveno.core.user.DontShowAgainLookup;
@@ -141,6 +143,8 @@ import java.net.URISyntaxException;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Paths;
+import java.text.DecimalFormat;
+import java.text.NumberFormat;
 import java.util.ArrayList;
 import java.util.Currency;
 import java.util.HashMap;
@@ -629,6 +633,19 @@ public class GUIUtil {
                 }
             }
         };
+    }
+
+    // the amount approximated in the user's preferred currency, or null while no price is available
+    public static String getFiatText(BigInteger atomicAmount, PriceFeedService priceFeedService, Preferences preferences) {
+        if (atomicAmount == null || atomicAmount.signum() <= 0 || priceFeedService == null || preferences == null) return null;
+        TradeCurrency currency = preferences.getPreferredTradeCurrency();
+        if (currency == null) return null;
+        MarketPrice price = priceFeedService.getMarketPrice(currency.getCode());
+        if (price == null || !price.isPriceAvailable()) return null;
+        DecimalFormat format = (DecimalFormat) NumberFormat.getNumberInstance(GlobalSettings.getLocale());
+        format.setMinimumFractionDigits(2);
+        format.setMaximumFractionDigits(2);
+        return "≈ " + format.format(HavenoUtils.atomicUnitsToXmr(atomicAmount) * price.getPrice()) + " " + currency.getCode();
     }
 
     public static void updateConfidence(MoneroTx tx,
